+++ frontend/src/pages/MemberData/Watchlist--Settings.vue
@@ -11,10 +11,8 @@
                 Alliances and corporations whose members are included in the list if they
                 also have characters in other (not NPC) corporations.
             </p>
-            <admin v-if="alliances" :sticky="sticky" :contentType="'alliances'"
-                   :type="'Watchlist'" :typeId="list.id" :selectContent="alliances"></admin>
-            <admin :sticky="sticky" :contentType="'corporations'"
-                   :type="'Watchlist'" :typeId="list.id"></admin>
+            <admin :sticky="sticky" :contentType="'alliances'" :type="'Watchlist'" :typeId="list.id"></admin>
+            <admin :sticky="sticky" :contentType="'corporations'" :type="'Watchlist'" :typeId="list.id"></admin>
         </div>
 
         <div class="card-header bg-light text-dark"><strong>Kicklist</strong></div>
@@ -33,8 +31,8 @@
                     <span class="far fa-plus-square"></span> corporations
                 </a>
             </p>
-            <admin ref="adminAllianceKick" v-if="alliances" :sticky="sticky" :contentType="'alliances'"
-                   :type="'WatchlistKicklist'" :typeId="list.id" :selectContent="alliances"></admin>
+            <admin ref="adminAllianceKick" :sticky="sticky" :contentType="'alliances'"
+                   :type="'WatchlistKicklist'" :typeId="list.id"></admin>
             <admin ref="adminCorpKick" :sticky="sticky" :contentType="'corporations'"
                    :type="'WatchlistKicklist'" :typeId="list.id"></admin>
         </div>
@@ -45,8 +43,8 @@
                 Alliances and corporations that should be treated like NPC corporations
                 (usually personal alt corporations).
             </p>
-            <admin v-if="alliances" :sticky="sticky" :contentType="'alliances'"
-                   :type="'WatchlistAllowlist'" :typeId="list.id" :selectContent="alliances"></admin>
+            <admin :sticky="sticky" :contentType="'alliances'"
+                   :type="'WatchlistAllowlist'" :typeId="list.id"></admin>
             <admin :sticky="sticky" :contentType="'corporations'"
                    :type="'WatchlistAllowlist'" :typeId="list.id"></admin>
         </div>
@@ -57,7 +55,6 @@
 <script>
 import AddEntity from '../../components/EntityAdd.vue';
 import Admin     from '../../components/EntityRelationEdit.vue';
-import {AllianceApi} from "neucore-js-client";
 
 export default {
     components: {
@@ -72,17 +69,7 @@ export default {
     data() {
         return {
             sticky: 87,
-            alliances: null,
         }
-    },
-
-    mounted() {
-        new AllianceApi().all((error, data) => {
-            if (error) {
-                return;
-            }
-            this.alliances = data;
-        });
     },
 
     methods: {