+++ CHANGELOG.md
@@ -6,6 +6,7 @@ dd mm yyyy
 
 - **BC Break** The body message of the IP-based rate limit error response (status 429) slightly changed.
 - Allow roles tracking and watchlist to update characters and services from all accounts.
+- Added Ajax search for the alliance and corporation select boxes instead of initially populating them with all.
 - Small UI fixes and improvements.
 - Some refactoring.